dialog integrated to loggroup

+++ detonator/src/main/java/airhacks/detonator/cloudwatch/boundary/LogGroups.java
@@ -1,5 +1,6 @@
 package airhacks.detonator.cloudwatch.boundary;
 
+import airhacks.detonator.dialog.control.Dialog;
 import software.amazon.awssdk.services.cloudwatchlogs.CloudWatchLogsClient;
 import software.amazon.awssdk.services.cloudwatchlogs.model.DeleteLogGroupRequest;
 import software.amazon.awssdk.services.cloudwatchlogs.model.LogGroup;
@@ -7,22 +8,26 @@ import software.amazon.awssdk.services.cloudwatchlogs.model.LogGroup;
 public interface LogGroups {
 
     static void removeAllLogGroups() {
-        try(var client = CloudWatchLogsClient.create()){
+        try (var client = CloudWatchLogsClient.create()) {
             client
-            .describeLogGroupsPaginator()
-            .stream()
-            .flatMap(response-> response.logGroups().stream())
-            .map(LogGroup::logGroupName)
-            .peek(System.out::println)
-            .forEach(groupName -> LogGroups.delete(client,groupName));
+                    .describeLogGroupsPaginator()
+                    .stream()
+                    .flatMap(response -> response.logGroups().stream())
+                    .map(LogGroup::logGroupName)
+                    .peek(System.out::println)
+                    .forEach(groupName -> LogGroups.delete(client, groupName));
 
         }
     }
-    
-    static void delete(CloudWatchLogsClient client,String logGroupName){
+
+    static void delete(CloudWatchLogsClient client, String logGroupName) {
         var deleteRequest = DeleteLogGroupRequest.builder()
-        .logGroupName(logGroupName)
-        .build();
-        client.deleteLogGroup(deleteRequest);
+                .logGroupName(logGroupName)
+                .build();
+        if (Dialog.proceed(logGroupName)) {
+            client.deleteLogGroup(deleteRequest);
+        }else{
+            
+        }
     }
 }
